requirements.txt
fastapi[standard]
SQLAlchemy
psycopg2-binary
bcrypt
cloudinary
python-dotenv
